Add mesh reader and writer
